requirements/base.txt
attrs==22.2.0
colorama==0.4.6
exceptiongroup==1.1.0
iniconfig==2.0.0
opencv-python==4.7.0.68
opyoid==1.7.0
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pyscreenshot==3.0
pytesseract==0.3.10
tomli==2.0.1
pytesseract==0.3.10
